LPS-160941 Move widgets to state, add thunks, actions and reducer
