+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/types.js
@@ -36,6 +36,7 @@ export const SWITCH_VIEWPORT_SIZE = 'SWITCH_VIEWPORT_SIZE';
 export const TOGGLE_FRAGMENT_HIGHLIGHTED = 'TOGGLE_FRAGMENT_HIGHLIGHTED';
 export const TOGGLE_PERMISSION = 'TOGGLE_PERMISSION';
 export const TOGGLE_SHOW_RESOLVED_COMMENTS = 'TOGGLE_SHOW_RESOLVED_COMMENTS';
+export const TOGGLE_WIDGET_HIGHLIGHTED = 'TOGGLE_WIDGET_HIGHLIGHTED';
 export const UNLOAD_REDUCER = 'UNLOAD_REDUCER';
 export const UPDATE_COL_SIZE = 'UPDATE_COL_SIZE';
 export const UPDATE_COLLECTION_DISPLAY_COLLECTION =
@@ -56,3 +57,4 @@ export const UPDATE_ROW_COLUMNS = 'UPDATE_ROW_COLUMNS';
 export const UPDATE_UNDO_ACTIONS = 'UPDATE_UNDO_ACTIONS';
 export const UPDATE_PAGE_CONTENTS = 'UPDATE_PAGE_CONTENTS';
 export const UPDATE_PREVIEW_IMAGE = 'UPDATE_PREVIEW_IMAGE';
+export const UPDATE_WIDGETS = 'UPDATE_WIDGETS';
