--- modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/contexts/WidgetsContext.js
@@ -1,182 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-import React, {useCallback, useContext, useEffect, useState} from 'react';
-
-import selectSegmentsExperienceId from '../selectors/selectSegmentsExperienceId';
-import WidgetService from '../services/WidgetService';
-import {useDispatch, useSelector, useSelectorRef} from './StoreContext';
-
-const HIGHLIGHTED_CATEGORY_ID = 'root--category-highlighted';
-
-const DEFAULT_HIGHLIGHTED_CATEGORY = {
-	categories: [],
-	path: HIGHLIGHTED_CATEGORY_ID,
-	portlets: [],
-	title: Liferay.Language.get('highlighted'),
-};
-
-const WidgetsContext = React.createContext([]);
-
-export function useWidgets() {
-	const {widgets} = useContext(WidgetsContext);
-
-	return widgets;
-}
-
-export function useToggleWidgetHighlighted() {
-	const {setWidgets, widgets} = useContext(WidgetsContext);
-	const dispatch = useDispatch();
-
-	return useCallback(
-		(portletId, highlighted) => {
-			WidgetService.toggleWidgetHighlighted({
-				highlighted,
-				onNetworkStatus: dispatch,
-				portletId,
-			}).then(({highlightedPortlets}) => {
-				const nextWidgets = widgets.reduce((categories, category) => {
-					if (category.path !== HIGHLIGHTED_CATEGORY_ID) {
-						categories.push({
-							...category,
-							portlets: category.portlets.map((widget) =>
-								widget.portletId === portletId
-									? {...widget, highlighted}
-									: widget
-							),
-						});
-					}
-
-					return categories;
-				}, []);
-
-				if (highlightedPortlets.length) {
-					nextWidgets.unshift({
-						...DEFAULT_HIGHLIGHTED_CATEGORY,
-						portlets: highlightedPortlets,
-					});
-				}
-
-				setWidgets(nextWidgets);
-			});
-		},
-		[setWidgets, widgets, dispatch]
-	);
-}
-
-function normalizePortlets(portlets, fragmentEntryLinks) {
-	return portlets.map((portlet) => {
-		const normalizedPortlet = {
-			...portlet,
-			used: Object.values(fragmentEntryLinks).some(
-				({portletId}) => portlet.portletId === portletId
-			),
-		};
-
-		if (portlet.portletItems?.length) {
-			normalizedPortlet.portletItems = normalizePortlets(
-				portlet.portletItems,
-				fragmentEntryLinks
-			);
-		}
-
-		return normalizedPortlet;
-	});
-}
-
-function normalizeCategories(categories, fragmentEntryLinks) {
-	return categories.map((category) => {
-		const normalizedCategory = {
-			...category,
-			portlets: normalizePortlets(category.portlets, fragmentEntryLinks),
-		};
-
-		if (category.categories?.length) {
-			normalizedCategory.categories = normalizeCategories(
-				category.categories,
-				fragmentEntryLinks
-			);
-		}
-
-		return normalizedCategory;
-	});
-}
-
-export function WidgetsContextProvider({children}) {
-	const [widgets, setWidgets] = useState([]);
-
-	const fragmentEntryLinksIds = useSelector((state) => {
-		const nextSegmentsExperienceId = selectSegmentsExperienceId(state);
-
-		return Object.values(state.fragmentEntryLinks)
-			.filter(
-				({portletId, removed, ...fragmentEntryLink}) =>
-					portletId &&
-					!removed &&
-					fragmentEntryLink.segmentsExperienceId ===
-						nextSegmentsExperienceId
-			)
-			.map(({fragmentEntryLinkId}) => fragmentEntryLinkId)
-			.join(',');
-	});
-
-	const fragmentEntryLinksRef = useSelectorRef((state) => {
-		const nextSegmentsExperienceId = selectSegmentsExperienceId(state);
-
-		return Object.values(state.fragmentEntryLinks).filter(
-			({portletId, removed, ...fragmentEntryLink}) =>
-				portletId &&
-				!removed &&
-				fragmentEntryLink.segmentsExperienceId ===
-					nextSegmentsExperienceId
-		);
-	});
-
-	useEffect(() => {
-		WidgetService.getWidgets().then((categories) =>
-			setWidgets(
-				normalizeCategories(categories, fragmentEntryLinksRef.current)
-			)
-		);
-	}, [fragmentEntryLinksRef]);
-
-	useEffect(() => {
-		setWidgets((currentWidgets) =>
-			normalizeCategories(currentWidgets, fragmentEntryLinksRef.current)
-		);
-	}, [fragmentEntryLinksIds, fragmentEntryLinksRef]);
-
-	useEffect(() => {
-		const handler = Liferay.on('addPortletConfigurationTemplate', () => {
-			WidgetService.getWidgets().then((categories) =>
-				setWidgets(
-					normalizeCategories(
-						categories,
-						fragmentEntryLinksRef.current
-					)
-				)
-			);
-		});
-
-		return () => {
-			handler.detach();
-		};
-	}, [fragmentEntryLinksRef]);
-
-	return (
-		<WidgetsContext.Provider value={{setWidgets, widgets}}>
-			{children}
-		</WidgetsContext.Provider>
-	);
-}
