+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/thunks/loadWidgets.js
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import updateWidgets from '../actions/updateWidgets';
+import WidgetService from '../services/WidgetService';
+
+export default function loadWidgets({fragmentEntryLinks}) {
+	return (dispatch) => {
+		WidgetService.getWidgets().then((widgets) =>
+			dispatch(
+				updateWidgets({
+					fragmentEntryLinks,
+					widgets,
+				})
+			)
+		);
+	};
+}
